Trukket ut logikk i egne komponenter
