+++ src/søknader/barnetilsyn/steg/5-aktivitet/AktivitetSyk.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { useLokalIntlContext } from '../../../../context/LokalIntlContext';
+import { hentHTMLTekst } from '../../../../utils/teksthåndtering';
+import AlertStripeDokumentasjon from '../../../../components/AlertstripeDokumentasjon';
+import { Alert, Label } from '@navikt/ds-react';
+
+export const AktivitetSyk: React.FC = () => {
+  const intl = useLokalIntlContext();
+
+  return (
+    <>
+      <Alert variant={'info'} inline>
+        <Label as="p">{hentHTMLTekst('erDuIArbeid.alertsstripe-info', intl)}</Label>
+      </Alert>
+      <AlertStripeDokumentasjon>
+        {hentHTMLTekst('erDuIArbeid.alertsstripe-dokumentasjon', intl)}
+      </AlertStripeDokumentasjon>
+    </>
+  );
+};
